Move constants file to src
